backend/requirements.txt
fastapi
uvicorn
google-generativeai
pandas
python-dotenv